move opts, handle booleans as literals
